airflow_job/requirements.txt
pandas>=2.0.0
sqlalchemy>=1.4.0
apache-airflow-providers-postgres
dbt-postgres